X410 SPI: Dropped Transactions Workaround
The X410/UHD has a bug where SPI transactions are dropped if received in quick succession. This results in unreliable communication with the SC2430. NI/Ettus has been made aware of this issue.

As a short-term solution to ensure reliable communication, SPI transaction readbacks have been enabled to provide throttling. The consequence of this workaround is an increase in latency (~3 orders of magnitude) when performing successive SPI write transactions.